LayoutTests/inspector/runtime/evaluate-emulateUserGesture-expected.txt

Tests for Runtime emulateUserGesture option.


== Running test suite: Runtime.evaluate.emulateUserGesture
-- Running test case: EvaluateWithoutEmulatingUserGesture
Not in User Gesture
-- Running test case: EvaluateWithEmulatingUserGesture
In User Gesture
